% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummies.R
\name{dummies}
\alias{dummies}
\title{Convert a N-category vector to a N-dimension matrix}
\usage{
dummies(x)
}
\arguments{
\item{x}{A N-category vector}
}
\value{
A N-dimension matrix with 0/1 values
}
\description{
The function \code{dummies} converts a N-category vector to a N-dimension matrix
}
\examples{
data(iris, package = "datasets")
dummies(iris[, 5])
}
