% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{folds}
\alias{folds}
\title{Generate a list of index for the n-fold cross-validation}
\usage{
folds(idx, n, seed = 1)
}
\arguments{
\item{idx}{A vector of index list}

\item{n}{The number of n folds}

\item{seed}{The seed value to generate random n-fold index}
}
\value{
A list of n-fold index
}
\description{
The function \code{folds} generates a list of index for the n-fold cross-validation
}
\examples{
folds(seq(10), 3, 2020)
}
