% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_latin.R
\name{gen_latin}
\alias{gen_latin}
\title{Generate random numbers of latin hypercube sampling}
\usage{
gen_latin(min = 0, max = 1, n, seed = 1)
}
\arguments{
\item{min}{The minimum value of random numbers}

\item{max}{The maxinum value of random numbers}

\item{n}{The number of random numbers to gernate}

\item{seed}{The seed value of random number generation}
}
\value{
A vector of random numbers bounded by the min and max
}
\description{
The function \code{gen_latin} generates a vector of random numbers by latin hypercube sampling
}
\examples{
gen_latin(0, 1, 10, 2020)
}
