% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sobol.R
\name{gen_sobol}
\alias{gen_sobol}
\title{Generate sobol sequence}
\usage{
gen_sobol(min = 0, max = 1, n, seed = 1)
}
\arguments{
\item{min}{The minimum value of random numbers}

\item{max}{The maxinum value of random numbers}

\item{n}{The number of random numbers to gernate}

\item{seed}{The seed value of random number generation}
}
\value{
A vector of sobol sequence bounded by the min and max
}
\description{
The function \code{gen_sobol} generates a vector of scrambled sobol sequence
}
\examples{
gen_sobol(0, 1, 10, 2020)
}
